Attempt output current value from PIR sensor to GUI_4

+++ t_12/TouchGFX/gui/include/gui/main_screen/MainView.hpp
@@ -28,7 +28,7 @@ protected:
 
     // ______________ View to Presenter ______________
 
-    void userAction(int value);
+    void userAction(uint16_t value);
 
 private:
     uint8_t count;
